Fix #16: Refactor Modules
