--- src/app.js
@@ -1,1 +1,0 @@
-import Popup from './popup/popup.js'
