+++ src/background/background.html
@@ -5,12 +5,8 @@
 	<meta charset="UTF-8">
 	<title>Document</title>
   	<script src="https://www.gstatic.com/firebasejs/3.3.0/firebase.js"></script>
-	<script type="text/javascript" src="background.js"></script>
 
 </head>
 <body>
-	
-
-<script type="text/javascript" src="background.js"></script>
 </body>
 </html>
